separate scheduler_job unit tests into different modules for different subclasses
