+++ cstar/tests/unit_tests/system/test_scheduler_job.py
@@ -1,43 +1,15 @@
-# 148, 154, 158-174, 194, 198, 219-228, 241-242, 358, 369, 416, 434, 451-456, 478, 498
-# 195, 199, 220-229, 242-243, 359, 370, 417, 435, 452-457, 479, 499
-# 359, 370, 417, 435, 452-457, 479, 499
-# TestSchedulerJobBase
-# - Test of warning when queue.max_walltime is missing [111]
-# - Test of different pathways to nodes and cpus_per_node [entire block, 158]
-# - Test of confirmation that user wants to continue with updates(seconds==0) [219]
-# - Test of catching KeyboardInterrupt in updates() [241]
-
-# TestSlurmJob
-# - Test of raise when sbatch return code is nonzero [358]
-# - Test of raise when no matches for JobID regex [369]
-
-# TestPBSJob
-# - case of "unsubmitted" in status property when id is None [416]
-# - Test of raise in status property on inability to find JobID in qstat/JSON fields [434]
-# - Check for "E" in status property [451]
-# - Check of unknown job_state in status property [454]
-# - Test of raise in status property for JSONDecodeError [456]
-# - Test of raise in submit method when regex can't find JobID [478]
-# - Test of raise in cancel method when returncode is 0
-
-
-import json
 import time
 import pytest
 import threading
 from unittest.mock import MagicMock, patch, PropertyMock
 from cstar.system.scheduler import (
     Scheduler,
-    SlurmScheduler,
     PBSScheduler,
-    SlurmQueue,
     PBSQueue,
 )
 from cstar.system.scheduler_job import (
     SchedulerJob,
-    SlurmJob,
     PBSJob,
-    create_scheduler_job,
 )
 
 
@@ -405,759 +377,3 @@ class TestCalculateNodeDistribution:
 
 
 ##
-
-
-class TestSlurmJob:
-    """Tests for the SlurmJob class."""
-
-    def setup_method(self, method):
-        # Mock the environment variables
-        self.mock_env_vars = {
-            "MPIHOME": "/mock/mpi",
-            "NETCDFHOME": "/mock/netcdf",
-            "LD_LIBRARY_PATH": "/mock/lib",
-        }
-
-        # Define common job parameters
-        self.common_job_params = {
-            "scheduler": MockScheduler(),
-            "commands": "echo Hello, World",
-            "account_key": "test_account",
-            "cpus": 4,
-            "walltime": "01:00:00",
-            "job_name": "test_job",
-        }
-
-    @patch("cstar.system.manager.CStarSystemManager.name", new_callable=PropertyMock)
-    @patch(
-        "cstar.system.manager.CStarSystemManager.environment", new_callable=PropertyMock
-    )
-    def test_script(self, mock_environment, mock_sysname, tmp_path):
-        # Mock environment
-        mock_environment.return_value.environment_variables = self.mock_env_vars
-        mock_environment.return_value.uses_lmod = True
-        mock_sysname.return_value = "mock_system"
-
-        # Mock package_root to point to tmp_path
-        mock_environment.return_value.package_root = tmp_path
-
-        # Create the necessary mock .lmod file in the mocked package_root
-        lmod_dir = tmp_path / "additional_files/lmod_lists"
-        lmod_dir.mkdir(parents=True, exist_ok=True)
-        lmod_file = lmod_dir / "mock_system.lmod"
-        lmod_modules = ["moduleA", "moduleB"]
-        lmod_file.write_text("\n".join(lmod_modules))
-
-        # Initialize the job
-        job = SlurmJob(
-            **self.common_job_params,
-            output_file="/test/output.log",
-            queue_name="test_queue",
-        )
-
-        expected_script = (
-            "#!/bin/bash\n"
-            "#SBATCH --job-name=test_job\n"
-            "#SBATCH --output=/test/output.log\n"
-            "#SBATCH --mock_flag=test_queue\n"
-            "#SBATCH --ntasks=4\n"
-            "#SBATCH --account=test_account\n"
-            "#SBATCH --export=NONE\n"
-            "#SBATCH --mail-type=ALL\n"
-            "#SBATCH --time=01:00:00\n"
-            "#SBATCH -mock_directive mock_value"
-            "\nmodule reset\n"
-            "module load moduleA\n\n"
-            "module load moduleB"
-            "\nprintenv\n"
-            'export MPIHOME="/mock/mpi"\n'
-            'export NETCDFHOME="/mock/netcdf"\n'
-            'export LD_LIBRARY_PATH="/mock/lib"\n'
-            "\necho Hello, World"
-        )
-
-        assert (
-            job.script.strip() == expected_script.strip()
-        ), f"Expected:\n{expected_script}\n\nGot:\n{job.script}"
-
-    @patch("subprocess.run")
-    @patch(
-        "cstar.system.manager.CStarSystemManager.environment", new_callable=PropertyMock
-    )
-    @patch.dict(
-        "os.environ",
-        {"SLURM_JOB_ID": "123", "SLURM_NODELIST": "mock_node", "SOME_ENV_VAR": "value"},
-        clear=True,
-    )
-    def test_submit(self, mock_environment, mock_subprocess, tmp_path):
-        # Mock environment
-        mock_environment.return_value.environment_variables = self.mock_env_vars
-        mock_environment.return_value.uses_lmod = False
-
-        # Create temporary directory for the job script
-        script_path = tmp_path / "test_job.sh"
-        run_path = tmp_path
-
-        # Mock the subprocess.run behavior for sbatch
-        mock_subprocess.return_value = MagicMock(
-            returncode=0, stdout="Submitted batch job 12345\n", stderr=""
-        )
-
-        # Initialize the job
-        job = SlurmJob(
-            **self.common_job_params,
-            script_path=script_path,
-            run_path=run_path,
-        )
-
-        job.submit()
-
-        # Check that the script file is saved and matches job.script
-        assert script_path.exists()
-        with script_path.open() as f:
-            script_content = f.read()
-        assert (
-            script_content.strip() == job.script.strip()
-        ), f"Script content mismatch!\nExpected:\n{job.script}\nGot:\n{script_content}"
-
-        # Hardcoded expected environment (excluding SLURM_ variables)
-        expected_env = {"SOME_ENV_VAR": "value"}
-
-        # Check that sbatch was called correctly
-        mock_subprocess.assert_called_once_with(
-            f"sbatch {script_path}",
-            shell=True,
-            cwd=run_path,
-            env=expected_env,
-            capture_output=True,
-            text=True,
-        )
-
-        # Check that the job ID was set
-        assert job.id == 12345
-
-    @pytest.mark.parametrize(
-        "subprocess_stdout, subprocess_returncode, expected_exception_message",
-        [
-            # Case 1: Non-zero return code from subprocess.run
-            ("", 1, "Non-zero exit code when submitting job. STDERR:"),
-            # Case 2: Missing job ID in subprocess stdout
-            (
-                "Submitted job without ID",
-                0,
-                "Failed to parse job ID from sbatch output",
-            ),
-        ],
-    )
-    @patch("subprocess.run")
-    def test_submit_raises(
-        self,
-        mock_subprocess,
-        tmp_path,
-        subprocess_stdout,
-        subprocess_returncode,
-        expected_exception_message,
-    ):
-        # Mock the subprocess.run behavior
-        mock_subprocess.return_value = MagicMock(
-            returncode=subprocess_returncode,
-            stdout=subprocess_stdout,
-            stderr="Error" if subprocess_returncode != 0 else "",
-        )
-
-        # Create temporary directory for the job script
-        script_path = tmp_path / "test_job.sh"
-        run_path = tmp_path
-
-        # Initialize the job
-        job = SlurmJob(
-            **self.common_job_params,
-            script_path=script_path,
-            run_path=run_path,
-        )
-
-        # Assert that the correct exception is raised with the expected message
-        with pytest.raises(RuntimeError, match=expected_exception_message):
-            job.submit()
-
-    @patch("subprocess.run")
-    def test_cancel(self, mock_subprocess, tmp_path):
-        # Create a temporary directory for the job run path
-        run_path = tmp_path
-
-        # Mock the subprocess.run behavior for successful cancellation
-        mock_subprocess.return_value = MagicMock(returncode=0, stdout="", stderr="")
-
-        # Initialize the job with a set job ID
-        job = SlurmJob(
-            **self.common_job_params,
-            run_path=run_path,
-        )
-        job._id = 12345  # Manually set the job ID
-
-        # Call cancel
-        job.cancel()
-
-        # Check that scancel was called correctly
-        mock_subprocess.assert_called_once_with(
-            "scancel 12345",
-            shell=True,
-            cwd=run_path,
-            capture_output=True,
-            text=True,
-        )
-
-        # Reset the mock for the next scenario
-        mock_subprocess.reset_mock()
-
-        # Mock the subprocess.run behavior for failed cancellation
-        mock_subprocess.return_value = MagicMock(
-            returncode=1, stdout="", stderr="Error: Job not found"
-        )
-
-        # Call cancel and expect a RuntimeError
-        with pytest.raises(
-            RuntimeError, match="Non-zero exit code when cancelling job."
-        ):
-            job.cancel()
-
-        # Verify that scancel was still called
-        mock_subprocess.assert_called_once_with(
-            "scancel 12345",
-            shell=True,
-            cwd=run_path,
-            capture_output=True,
-            text=True,
-        )
-
-    def test_save_script(self, tmp_path):
-        # Define paths for the script file
-        script_path = tmp_path / "test_job.sh"
-
-        # Initialize the job
-        job = SlurmJob(
-            **self.common_job_params,
-            script_path=script_path,
-        )
-
-        # Call save_script
-        job.save_script()
-
-        # Check that the file was created
-        assert script_path.exists(), "The script file was not created."
-
-        # Verify the file content matches job.script
-        with script_path.open() as f:
-            file_content = f.read()
-
-        assert (
-            file_content.strip() == job.script.strip()
-        ), f"Script content mismatch!\nExpected:\n{job.script}\nGot:\n{file_content}"
-
-    @patch("subprocess.run")
-    @pytest.mark.parametrize(
-        "job_id, sacct_output, return_code, expected_status, should_raise",
-        [
-            (None, "", 0, "unsubmitted", False),  # Unsubmitted job
-            (12345, "PENDING\n", 0, "pending", False),  # Pending job
-            (12345, "RUNNING\n", 0, "running", False),  # Running job
-            (12345, "COMPLETED\n", 0, "completed", False),  # Completed job
-            (12345, "CANCELLED\n", 0, "cancelled", False),  # Cancelled job
-            (12345, "FAILED\n", 0, "failed", False),  # Failed job
-            (12345, "", 1, None, True),  # sacct command failure
-        ],
-    )
-    def test_status(
-        self,
-        mock_subprocess,
-        job_id,
-        sacct_output,
-        return_code,
-        expected_status,
-        should_raise,
-    ):
-        # Initialize the job
-        job = SlurmJob(**self.common_job_params)
-
-        # Set the job ID if provided
-        if job_id is not None:
-            job._id = job_id
-
-        # Mock subprocess.run behavior
-        mock_subprocess.return_value = MagicMock(
-            returncode=return_code,
-            stdout=sacct_output,
-            stderr="Error: sacct command failed" if return_code != 0 else "",
-        )
-
-        # Check the expected outcome
-        if should_raise:
-            with pytest.raises(
-                RuntimeError, match="Failed to retrieve job status using sacct"
-            ):
-                job.status
-        else:
-            assert (
-                job.status == expected_status
-            ), f"Expected status '{expected_status}' but got '{job.status}'"
-
-
-class TestPBSJob:
-    """Tests for the PBSJob class."""
-
-    def setup_method(self, method):
-        # Shared setup for PBSJob tests
-        self.common_job_params = {
-            "scheduler": MockScheduler(),
-            "commands": "echo Hello, World",
-            "account_key": "test_account",
-            "cpus": 4,
-            "nodes": 2,
-            "cpus_per_node": 2,
-            "walltime": "02:00:00",
-            "job_name": "test_pbs_job",
-            "output_file": "/test/pbs_output.log",
-            "queue_name": "test_queue",
-        }
-
-    @patch(
-        "cstar.system.manager.CStarSystemManager.scheduler", new_callable=PropertyMock
-    )
-    def test_script(self, mock_scheduler):
-        # Mock the scheduler and its attributes
-        mock_scheduler.return_value = MagicMock()
-        mock_scheduler.return_value.queue_flag = "q"
-        mock_scheduler.return_value.other_scheduler_directives = {
-            "mock_directive": "mock_value"
-        }
-
-        # Initialize a PBSJob
-        job = PBSJob(**self.common_job_params)
-
-        expected_script = (
-            "#PBS -S /bin/bash\n"
-            "#PBS -N test_pbs_job\n"
-            "#PBS -o /test/pbs_output.log\n"
-            "#PBS -A test_account\n"
-            "#PBS -l select=2:ncpus=2,walltime=02:00:00\n"
-            "#PBS -mock_flag test_queue\n"
-            "#PBS -j oe\n"
-            "#PBS -k eod\n"
-            "#PBS -V\n"
-            "#PBS mock_directive mock_value"
-            "\ncd ${PBS_O_WORKDIR}"
-            "\n\necho Hello, World"
-        )
-
-        # Validate the script content
-        assert (
-            job.script.strip() == expected_script.strip()
-        ), f"Script mismatch!\nExpected:\n{expected_script}\n\nGot:\n{job.script}"
-
-    @patch("subprocess.run")
-    @patch(
-        "cstar.system.manager.CStarSystemManager.scheduler", new_callable=PropertyMock
-    )
-    def test_submit(self, mock_scheduler, mock_subprocess, tmp_path):
-        # Mock scheduler attributes
-        mock_scheduler.return_value = MagicMock()
-        mock_scheduler.return_value.queue_flag = "q"
-        mock_scheduler.return_value.other_scheduler_directives = {}
-
-        # Mock subprocess.run for qsub
-        mock_subprocess.return_value = MagicMock(
-            returncode=0, stdout="12345.mockserver\n", stderr=""
-        )
-
-        # Create temporary paths
-        script_path = tmp_path / "test_pbs_job.sh"
-        run_path = tmp_path
-
-        # Initialize a PBSJob
-        job = PBSJob(
-            **self.common_job_params,
-            script_path=script_path,
-            run_path=run_path,
-        )
-
-        # Submit the job
-        job.submit()
-
-    @pytest.mark.parametrize(
-        "returncode, stdout, stderr, match_message",
-        [
-            (1, "", "Error submitting job", "Non-zero exit code when submitting job"),
-            (0, "InvalidJobIDFormat", "", "Unexpected job ID format from qsub"),
-        ],
-    )
-    @patch("subprocess.run")
-    @patch(
-        "cstar.system.manager.CStarSystemManager.scheduler", new_callable=PropertyMock
-    )
-    def test_submit_raises(
-        self,
-        mock_scheduler,
-        mock_subprocess,
-        returncode,
-        stdout,
-        stderr,
-        match_message,
-        tmp_path,
-    ):
-        # Mock scheduler attributes
-        mock_scheduler.return_value = MagicMock()
-        mock_scheduler.return_value.queue_flag = "q"
-        mock_scheduler.return_value.other_scheduler_directives = {}
-
-        # Mock subprocess.run for qsub
-        mock_subprocess.return_value = MagicMock(
-            returncode=returncode, stdout=stdout, stderr=stderr
-        )
-
-        # Create temporary paths
-        script_path = tmp_path / "test_pbs_job.sh"
-        run_path = tmp_path
-
-        # Initialize a PBSJob
-        job = PBSJob(
-            **self.common_job_params,
-            script_path=script_path,
-            run_path=run_path,
-        )
-
-        # Check for expected exceptions
-        with pytest.raises(RuntimeError, match=match_message):
-            job.submit()
-
-    @patch("subprocess.run")
-    @patch("cstar.system.scheduler_job.PBSJob.status", new_callable=PropertyMock)
-    def test_cancel_running_job(self, mock_status, mock_subprocess, tmp_path):
-        # Mock the status to "running"
-        mock_status.return_value = "running"
-
-        # Mock subprocess.run for successful cancellation
-        mock_subprocess.return_value = MagicMock(returncode=0, stdout="", stderr="")
-
-        # Create a PBSJob with a set job ID
-        job = PBSJob(
-            **self.common_job_params,
-            run_path=tmp_path,
-        )
-        job._id = 12345  # Manually set the job ID
-
-        # Cancel the job
-        job.cancel()
-
-        # Verify qdel was called correctly
-        mock_subprocess.assert_called_once_with(
-            "qdel 12345",
-            shell=True,
-            cwd=tmp_path,
-            capture_output=True,
-            text=True,
-        )
-
-    @patch("subprocess.run")
-    @patch("cstar.system.scheduler_job.PBSJob.status", new_callable=PropertyMock)
-    @patch("builtins.print")
-    def test_cancel_completed_job(
-        self, mock_print, mock_status, mock_subprocess, tmp_path
-    ):
-        # Mock the status to "completed"
-        mock_status.return_value = "completed"
-
-        # Create a PBSJob with a set job ID
-        job = PBSJob(
-            **self.common_job_params,
-            run_path=tmp_path,
-        )
-        job._id = 12345  # Manually set the job ID
-
-        # Attempt to cancel the job
-        job.cancel()
-
-        # Verify the message was printed
-        mock_print.assert_called_with("Cannot cancel job with status completed")
-
-        # Verify qdel was not called
-        mock_subprocess.assert_not_called()
-
-    ##
-    @patch("subprocess.run")
-    @patch("cstar.system.scheduler_job.PBSJob.status", new_callable=PropertyMock)
-    def test_cancel_failure(self, mock_status, mock_subprocess, tmp_path):
-        # Mock the status to "running"
-        mock_status.return_value = "running"
-
-        # Mock subprocess.run for qdel failure
-        mock_subprocess.return_value = MagicMock(
-            returncode=1, stdout="", stderr="Error cancelling job"
-        )
-
-        # Create a PBSJob with a set job ID
-        job = PBSJob(
-            **self.common_job_params,
-            run_path=tmp_path,
-        )
-        job._id = 12345  # Manually set the job ID
-
-        # Expect an error when qdel fails
-        with pytest.raises(
-            RuntimeError, match="Non-zero exit code when cancelling job"
-        ):
-            job.cancel()
-
-    ##
-    @pytest.mark.parametrize(
-        "qstat_output, exit_status, expected_status, should_raise, expected_exception, expected_message",
-        [
-            (
-                {"Jobs": {"12345": {"job_state": "Q"}}},
-                None,
-                "pending",
-                False,
-                None,
-                None,
-            ),  # Pending job
-            (
-                {"Jobs": {"12345": {"job_state": "R"}}},
-                None,
-                "running",
-                False,
-                None,
-                None,
-            ),  # Running job
-            (
-                {"Jobs": {"12345": {"job_state": "C"}}},
-                None,
-                "completed",
-                False,
-                None,
-                None,
-            ),  # Completed job
-            (
-                {"Jobs": {"12345": {"job_state": "H"}}},
-                None,
-                "held",
-                False,
-                None,
-                None,
-            ),  # Held job
-            (
-                {"Jobs": {"12345": {"job_state": "F", "Exit_status": 1}}},
-                None,
-                "failed",
-                False,
-                None,
-                None,
-            ),  # Failed job
-            (
-                {"Jobs": {"12345": {"job_state": "F", "Exit_status": 0}}},
-                None,
-                "completed",
-                False,
-                None,
-                None,
-            ),  # Completed with Exit_status 0
-            (
-                None,
-                1,
-                None,
-                True,
-                RuntimeError,
-                "Failed to retrieve job status using qstat",
-            ),  # qstat command failure
-            (
-                {"Jobs": {}},
-                None,
-                None,
-                True,
-                RuntimeError,
-                "Job ID 12345 not found in qstat output.",
-            ),  # Missing job info
-            (
-                {"Jobs": {"12345": {"job_state": "E"}}},
-                None,
-                "ending",
-                False,
-                None,
-                None,
-            ),  # Ending job
-            (
-                {"Jobs": {"12345": {"job_state": "X"}}},
-                None,
-                "unknown (X)",
-                False,
-                None,
-                None,
-            ),  # Unknown job state
-            (
-                "invalid_json",
-                None,
-                None,
-                True,
-                RuntimeError,
-                "Failed to parse JSON from qstat output",
-            ),  # JSONDecodeError
-        ],
-    )
-    @patch("subprocess.run")
-    def test_status(
-        self,
-        mock_subprocess,
-        qstat_output,
-        exit_status,
-        expected_status,
-        should_raise,
-        expected_exception,
-        expected_message,
-    ):
-        # Mock qstat command output
-        if qstat_output is not None:
-            if qstat_output == "invalid_json":
-                mock_subprocess.return_value = MagicMock(
-                    returncode=0, stdout="Invalid JSON", stderr=""
-                )
-            else:
-                mock_subprocess.return_value = MagicMock(
-                    returncode=0, stdout=json.dumps(qstat_output), stderr=""
-                )
-        else:
-            mock_subprocess.return_value = MagicMock(
-                returncode=1, stdout="", stderr="Error: qstat command failed"
-            )
-
-        # Create a PBSJob with a set job ID
-        job = PBSJob(**self.common_job_params)
-        job._id = 12345  # Manually set the job ID
-
-        # Check the expected outcome
-        if should_raise:
-            with pytest.raises(expected_exception, match=expected_message):
-                job.status
-        else:
-            assert (
-                job.status == expected_status
-            ), f"Expected status '{expected_status}' but got '{job.status}'"
-
-    def test_status_unsubmitted(self):
-        # Create a PBSJob without setting a job ID
-        job = PBSJob(**self.common_job_params)
-
-        # Assert the status is "unsubmitted"
-        assert (
-            job.status == "unsubmitted"
-        ), "Expected 'unsubmitted' status when job ID is None"
-
-    @patch("json.loads", side_effect=json.JSONDecodeError("Expecting value", "", 0))
-    @patch("subprocess.run")
-    def test_status_json_decode_error(self, mock_subprocess, mock_json_loads):
-        # Mock qstat command output
-        mock_subprocess.return_value = MagicMock(
-            returncode=0, stdout="Invalid JSON", stderr=""
-        )
-
-        # Create a PBSJob with a set job ID
-        job = PBSJob(**self.common_job_params)
-        job._id = 12345  # Manually set the job ID
-
-        # Check for JSONDecodeError handling
-        with pytest.raises(
-            RuntimeError, match="Failed to parse JSON from qstat output"
-        ):
-            job.status
-
-
-class TestCreateSchedulerJob:
-    @patch(
-        "cstar.system.manager.CStarSystemManager.scheduler", new_callable=PropertyMock
-    )
-    @patch("cstar.system.scheduler.SlurmQueue.max_walltime", new_callable=PropertyMock)
-    def test_create_slurm_job(self, mock_max_walltime, mock_scheduler):
-        # Mock max_walltime for the queue
-        mock_max_walltime.return_value = "02:00:00"
-
-        # Mock the scheduler to be a SlurmScheduler with a valid queue
-        mock_queue = SlurmQueue(name="test_queue", query_name="test_queue")
-        mock_scheduler.return_value = SlurmScheduler(
-            queues=[mock_queue],
-            primary_queue_name="test_queue",
-            queue_flag="mock_flag",
-        )
-
-        # Explicitly provide `queue_name`
-        job = create_scheduler_job(
-            commands="echo Hello, World",
-            cpus=4,
-            account_key="test_account",
-            walltime="01:00:00",
-            queue_name="test_queue",  # Explicitly specify queue_name
-        )
-
-        # Ensure the returned job is a SlurmJob instance
-        assert isinstance(job, SlurmJob), f"Expected SlurmJob, got {type(job).__name__}"
-        assert job.commands == "echo Hello, World"
-        assert job.cpus == 4
-        assert job.account_key == "test_account"
-        assert job.walltime == "01:00:00"  # Ensure the provided walltime is used
-
-    @patch(
-        "cstar.system.manager.CStarSystemManager.scheduler", new_callable=PropertyMock
-    )
-    @patch(
-        "cstar.system.scheduler.PBSScheduler.global_max_cpus_per_node",
-        new_callable=PropertyMock,
-    )
-    def test_create_pbs_job(self, mock_global_max_cpus, mock_scheduler):
-        # Mock global_max_cpus_per_node for the scheduler
-        mock_global_max_cpus.return_value = 128
-
-        # Mock the scheduler to be a PBSScheduler with a valid queue
-        mock_queue = PBSQueue(name="test_queue", max_walltime="02:00:00")
-        mock_scheduler.return_value = PBSScheduler(
-            queues=[mock_queue],
-            primary_queue_name="test_queue",
-            queue_flag="mock_flag",
-        )
-
-        # Explicitly provide `queue_name`
-        job = create_scheduler_job(
-            commands="echo Hello, World",
-            cpus=8,
-            account_key="pbs_account",
-            walltime="02:00:00",
-            nodes=1,
-            cpus_per_node=8,
-            queue_name="test_queue",  # Explicitly specify queue_name
-        )
-
-        # Ensure the returned job is a PBSJob instance
-        assert isinstance(job, PBSJob), f"Expected PBSJob, got {type(job).__name__}"
-        assert job.commands == "echo Hello, World"
-        assert job.cpus == 8
-        assert job.account_key == "pbs_account"
-        assert job.walltime == "02:00:00"  # Ensure the provided walltime is used
-
-    @patch(
-        "cstar.system.manager.CStarSystemManager.scheduler", new_callable=PropertyMock
-    )
-    def test_unsupported_scheduler(self, mock_scheduler):
-        # Mock an unsupported scheduler type
-        mock_scheduler.return_value = None  # No scheduler set
-
-        with pytest.raises(TypeError, match="Unsupported scheduler type: NoneType"):
-            create_scheduler_job(
-                commands="echo Hello, World",
-                cpus=4,
-                account_key="test_account",
-                walltime="01:00:00",
-            )
-
-    def test_missing_arguments(self):
-        # No need to mock scheduler here as the error is in the function arguments
-        with pytest.raises(TypeError, match="missing .* required positional argument"):
-            create_scheduler_job(
-                cpus=4,
-                account_key="test_account",
-                walltime="01:00:00",
-            )
